Move rule node converters to feature module
