+++ features/sharding/core/src/main/java/org/apache/shardingsphere/sharding/yaml/swapper/NewYamlShardingRuleConfigurationSwapper.java
@@ -22,7 +22,7 @@ import org.apache.shardingsphere.infra.util.yaml.YamlEngine;
 import org.apache.shardingsphere.infra.util.yaml.datanode.YamlDataNode;
 import org.apache.shardingsphere.infra.yaml.config.swapper.algorithm.YamlAlgorithmConfigurationSwapper;
 import org.apache.shardingsphere.infra.yaml.config.swapper.rule.NewYamlRuleConfigurationSwapper;
-import org.apache.shardingsphere.metadata.persist.node.metadata.config.sharding.ShardingNodeConverter;
+import org.apache.shardingsphere.sharding.metadata.converter.ShardingNodeConverter;
 import org.apache.shardingsphere.sharding.api.config.ShardingRuleConfiguration;
 import org.apache.shardingsphere.sharding.api.config.rule.ShardingAutoTableRuleConfiguration;
 import org.apache.shardingsphere.sharding.api.config.rule.ShardingTableReferenceRuleConfiguration;
@@ -40,7 +40,7 @@ import java.util.LinkedHashSet;
 import java.util.Map.Entry;
 
 /**
- * TODO Rename to NewYamlShardingRuleConfigurationSwapper when metadata structure adjustment completed. #25485
+ * TODO Rename to YamlShardingRuleConfigurationSwapper when metadata structure adjustment completed. #25485
  * New YAML sharding rule configuration swapper.
  */
 public final class NewYamlShardingRuleConfigurationSwapper implements NewYamlRuleConfigurationSwapper<ShardingRuleConfiguration> {
@@ -109,13 +109,13 @@ public final class NewYamlShardingRuleConfigurationSwapper implements NewYamlRul
     
     private void swapAlgorithms(final ShardingRuleConfiguration data, final Collection<YamlDataNode> result) {
         for (Entry<String, AlgorithmConfiguration> each : data.getShardingAlgorithms().entrySet()) {
-            result.add(new YamlDataNode(converter.getShardingAlgorithmsPath(each.getKey()), YamlEngine.marshal(each.getValue())));
+            result.add(new YamlDataNode(converter.getShardingAlgorithmPath(each.getKey()), YamlEngine.marshal(each.getValue())));
         }
         for (Entry<String, AlgorithmConfiguration> each : data.getKeyGenerators().entrySet()) {
-            result.add(new YamlDataNode(converter.getKeyGeneratorsPath(each.getKey()), YamlEngine.marshal(each.getValue())));
+            result.add(new YamlDataNode(converter.getKeyGeneratorPath(each.getKey()), YamlEngine.marshal(each.getValue())));
         }
         for (Entry<String, AlgorithmConfiguration> each : data.getAuditors().entrySet()) {
-            result.add(new YamlDataNode(converter.getAuditorsPath(each.getKey()), YamlEngine.marshal(each.getValue())));
+            result.add(new YamlDataNode(converter.getAuditorPath(each.getKey()), YamlEngine.marshal(each.getValue())));
         }
     }
