+++ kernel/metadata/core/src/main/java/org/apache/shardingsphere/metadata/persist/node/metadata/config/readwritesplitting/ReadwriteSplittingNodeConverter.java
@@ -22,25 +22,27 @@ package org.apache.shardingsphere.metadata.persist.node.metadata.config.readwrit
  */
 public final class ReadwriteSplittingNodeConverter {
     
+    private static final String DATA_SOURCES = "data_sources";
+    
     private static final String LOAD_BALANCER_NODE = "load_balancers";
     
     /**
      * Get group name path.
-     *
+     * 
      * @param groupName group name
      * @return group name path
      */
     public String getGroupNamePath(final String groupName) {
-        return String.join("/", "", groupName);
+        return String.join("/", DATA_SOURCES, groupName);
     }
     
     /**
      * Get load balancer name.
-     *
+     * 
      * @param loadBalancerName load balancer name
      * @return load balancer path
      */
     public String getLoadBalancerPath(final String loadBalancerName) {
-        return String.join("/", "", LOAD_BALANCER_NODE, loadBalancerName);
+        return String.join("/", LOAD_BALANCER_NODE, loadBalancerName);
     }
 }
